feat: Finished implementation
