+++ README.md
@@ -5,10 +5,15 @@
 - Git - [Download & Install Git](https://git-scm.com/downloads).
 - Node.js - [Download & Install Node.js](https://nodejs.org/en/download/) and the npm package manager.
 
-## Downloading
+## Downloading and checkout
 
 ```
-git clone {repository URL}
+git clone https://github.com/fructus0/nodejs2022Q4-service.git
+```
+
+After clone, run next command to get to actual code
+```
+git checkout development
 ```
 
 ## Installing NPM modules
@@ -19,11 +24,27 @@ npm install
 
 ## Running application
 
+***!!! Before running, you need to create a `.env` file and copy the contents of `.env.example` into it***
+
+___
+
+***Once this has been done, the following command will run the application and generate documentation:***
+
 ```
 npm start
 ```
 
-After starting the app on port (4000 as default) you can open
+## Developing
+
+To automate the restart of the application, when changing files - you should use the following command:
+
+```
+npm run start:dev
+```
+
+## Documentation
+
+After starting the app on port (4000 as default port in `.env`) you can open
 in your browser OpenAPI documentation by typing http://localhost:4000/doc/.
 For more information about OpenAPI/Swagger please visit https://swagger.io/.
 
@@ -64,9 +85,3 @@ npm run lint
 ```
 npm run format
 ```
-
-### Debugging in VSCode
-
-Press <kbd>F5</kbd> to debug.
-
-For more information, visit: https://code.visualstudio.com/docs/editor/debugging
