+++ src/artists/dto/update-artist.dto.ts
