--- src/dto/update-favorite.dto.ts
@@ -1,12 +1,0 @@
-import { IsUUID } from 'class-validator';
-import { UUID_VERSION } from '../constants/common';
-
-export class UpdateFavoriteDto {
-  @IsUUID(UUID_VERSION, { each: true })
-  artists: string[];
-  @IsUUID(UUID_VERSION, { each: true })
-  albums: string[];
-
-  @IsUUID(UUID_VERSION, { each: true })
-  tracks: string[];
-}
