+++ src/favorites/dto/get-favorites-response.dto.ts
@@ -1,0 +1,9 @@
+import { ArtistEntity } from '../../artists/entities/artist.entity';
+import { AlbumEntity } from '../../albums/entities/album.entity';
+import { TrackEntity } from '../../tracks/entities/track.entity';
+
+export class GetFavoritesResponseDto {
+  artists: ArtistEntity[];
+  albums: AlbumEntity[];
+  tracks: TrackEntity[];
+}
