+++ src/main.ts
@@ -1,9 +1,21 @@
+import { config } from 'dotenv';
+import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
+
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './modules/app.module';
+import { AppModule } from './app/app.module';
+
+config();
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  await app.listen(4000);
+  const config = new DocumentBuilder()
+    .setTitle('Nodejs2022Q4 Service')
+    .setVersion('1.0')
+    .build();
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('doc', app, document);
+
+  await app.listen(process.env.PORT);
 }
 
 bootstrap();
